backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.2
faiss-cpu==1.7.4